Add new badges to masonry view
